requirements.txt
requests
beautifulsoup4
langchain
transformers
torch
python-dotenv